bacula-sd/bacula-fd/bacula-director-common: Please add LSB formatted dependency info in init.d script
Package:  bacula-sd bacula-fd bacula-director-common
Version:  2.2.5-1
Tags:     patch
User:     initscripts-ng-devel@lists.alioth.debian.org
Usertags: missing-dependency

To be able to check boot script order, and also to be able to start
boot scripts in parallel, it is important to know the dependencies of
the various boot scripts.  The Linux Software Base specifies a init.d
header file format useful for this purpose, and adding such header to
the anacron init.d script would make it possible for me to use this
information to check the current sequence and to speed up the debian
boot.

I am working on a system to update the boot sequence based on these
dependencies, and would like see this as the default in Lenny.
Because of this, it is nice if the dependencies was updated quickly.

<URL:http://refspecs.freestandards.org/LSB_2.1.0/LSB-generic/LSB-generic/initscrcomconv.html>
documents the LSB header format.  Some debian notes are available from
<URL:http://wiki.debian.org/LSBInitScripts>.

Here is a patch to document the dependencies.  I hope this is correct.