Apply nobash.diff from Sami Haahtinen
From Sami:

nobash.diff addresses a problem when compiling on a system that has
dash as the default shell. If configure is explicitly run with bash
the shell variable tests will also be performed with bash while the
scripts are actually run with dash. (took me a while to track this
one)